+++ UltraPortal/Weapons/PortalGun.cs
@@ -146,10 +146,12 @@ namespace UltraPortal {
 
 		public void Reset() {
 			if (PortalEntry) {
+				PortalEntry.Reset();
 				PortalEntry.transform.position = DefaultPortalPosition;
 			}
 
 			if (PortalExit) {
+				PortalExit.Reset();
 				PortalExit.transform.position = DefaultPortalPosition;
 			}